requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
flake8==3.9.2
flake8-html==0.4.1
importlib-metadata==4.0.1
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
zipp==3.4.1
